feat(unit-test): start implementing unit test
