--- components/common/AirbnbDateRange/DateRange.module.css
@@ -1,22 +1,0 @@
-.container {
-  /* border: 2px solid red; */
-  background: #f9fafb !important;
-}
-
-.container .CalendarDay__selected_span {
-  background: #82e0aa !important;
-  color: white;
-  border: 1px solid blue;
-}
-.container .DateRangePickerInput__withBorder {
-  background: #f9fafb !important;
-}
-
-.container .DateInput {
-  width: 116px !important;
-  background: #f9fafb !important;
-}
-.container .DateInput_input {
-  background: #f9fafb !important;
-  font-size: 0.875rem !important;
-}
