+++ components/common/Footer.tsx
@@ -76,4 +76,4 @@ const Footer = () => (
 </div>
 );
 
-export default Footer;
+export {Footer};
