+++ components/common/Header.tsx
@@ -139,4 +139,4 @@ const Header = () => {
   )
 }
 
-export default Header
+export {Header}
