+++ components/index.js
@@ -1,22 +1,24 @@
-import Main from "./Main";
-import AirbnbDateRange from "./common/AirbnbDateRange";
-import CustomTableLoader from "./common/CustomTableLoader";
-import Footer from "./common/Footer";
-import Header from "./common/Header";
-import Modal from "./common/Modal";
-import Sidebar from "./common/Sidebar";
-import Tab from "./common/Tab";
-import Toaster, {
+
+
+import {CustomTableLoader} from "./common/CustomTableLoader";
+import {Footer} from "./common/Footer";
+import {Main} from "./common/Main";
+import {Header} from "./common/Header";
+import {Modal} from "./common/Modal";
+import {Sidebar} from "./common/Sidebar";
+import {Tab} from "./common/Tab";
+import  {
+  Toaster,
   showToastErrorMessage,
   showToastInfoMessage,
   showToastMessage,
   showToastWarningMessage,
 } from "./common/Toaster";
-import StudentActive from "./teacherAndStudent/StudentActive";
-import TeacherActive from "./teacherAndStudent/TeacherActive";
-import TeacherAndStudent from "./teacherAndStudent/TeacherAndStudent";
+import {StudentActive} from "./teacherAndStudent/StudentActive";
+import {TeacherActive} from "./teacherAndStudent/TeacherActive";
+import {TeacherAndStudent} from "./teacherAndStudent/TeacherAndStudent";
 
 export {
-  AirbnbDateRange, CustomTableLoader, Footer, Header, Main, Modal, Sidebar, StudentActive, Tab, TeacherActive, TeacherAndStudent, Toaster, showToastErrorMessage, showToastInfoMessage, showToastMessage, showToastWarningMessage
+   CustomTableLoader, Footer, Header, Modal, Sidebar, StudentActive, Tab, TeacherActive, TeacherAndStudent, Toaster, showToastErrorMessage, showToastInfoMessage, showToastMessage, showToastWarningMessage, Main
 };
